temp/tires_test_project/requirements.txt
Django==1.11.5
django-appconf==1.0.2
olefile==0.44
pilkit==2.0
Pillow==4.2.1
psycopg2==2.7.3.1
pytz==2017.2
six==1.11.0
